requirement.txt
absl-py==0.2.0
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.6.2
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
bleach==1.5.0
certifi==2018.8.24
cffi==1.11.5
cloudpickle==0.5.5
colorama==0.3.9
constantly==15.1.0
cryptography==2.3.1
cycler==0.10.0
dask==0.19.1
decorator==4.3.0
Django==1.11.15
django-filer==1.3.2
django-mptt==0.8.7
django-polymorphic==2.0.3
easy-thumbnails==2.5
entrypoints==0.2.3
gast==0.2.0
grpcio==1.11.0
html5lib==0.9999999
hyperlink==18.0.0
idna==2.7
imageio==2.4.1
incremental==17.5.0
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.8.3
mkl-fft==1.0.4
mkl-random==1.0.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.6.0
numpy==1.14.3
olefile==0.46
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
patsy==0.5.0
pickleshare==0.7.4
Pillow==5.2.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
pydicom==1.1.0
Pygments==2.2.0
pyOpenSSL==18.0.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.0
pywin32==223
pywinpty==0.5.4
pyzmq==17.1.2
qtconsole==4.4.1
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==1.1.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
sip==4.19.12
six==1.11.0
statsmodels==0.9.0
tensorboard==1.8.0
tensorflow==1.10.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tf-unet==0.1.1
toolz==0.9.0
tornado==5.1
traitlets==4.3.2
Twisted==18.7.0
Unidecode==0.4.21
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.1
wincertstore==0.2
zope.interface==4.5.0
